[rename] SearchButton 컴포넌트 파일명 오타 수정
